refactor(shared-charts): move playground
