+++ packages/pluggableWidgets/chart-playground-web/src/components/Sidebar.tsx
